requirements.txt
fastapi
uvicorn
gunicorn
sentence-transformers
faiss-cpu
numpy
python-dotenv
loguru
python-jose[cryptography]
passlib[bcrypt]
